+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ## Next version (TBA)
 
-*
+* Performance improvements
 
 
 ## 0.1.1 (2020-04-07)